Extract GUI opening
