+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/block/BlockListener.kt
@@ -15,7 +15,6 @@ import io.papermc.paper.event.player.PlayerLecternPageChangeEvent
 import io.papermc.paper.event.player.PlayerOpenSignEvent
 import org.bukkit.GameMode
 import org.bukkit.Material
-import org.bukkit.event.Event
 import org.bukkit.event.EventHandler
 import org.bukkit.event.EventPriority
 import org.bukkit.event.Listener
@@ -29,9 +28,6 @@ import org.bukkit.event.inventory.FurnaceBurnEvent
 import org.bukkit.event.inventory.FurnaceExtractEvent
 import org.bukkit.event.player.PlayerInteractEvent
 import org.bukkit.event.player.PlayerTakeLecternBookEvent
-import org.bukkit.inventory.EquipmentSlot
-import xyz.xenondevs.inventoryaccess.component.AdventureComponentWrapper
-import xyz.xenondevs.invui.window.Window
 
 
 /**
@@ -51,7 +47,7 @@ internal object BlockListener : Listener {
         val player = event.player
 
         val pylonItem = PylonItem.fromStack(item)
-        val pylonBlock = pylonItem?.schema?.place(BlockCreateContext.PlayerPlace(player, item, event.block))
+        val pylonBlock = pylonItem?.place(BlockCreateContext.PlayerPlace(player, item, event.block))
 
         if (pylonItem != null && pylonBlock == null) {
             event.isCancelled = true
@@ -487,21 +483,6 @@ internal object BlockListener : Listener {
         val pylonBlock = BlockStorage.get(event.clickedBlock ?: return)
         if (pylonBlock is PylonInteractableBlock) {
             pylonBlock.onInteract(event)
-        }
-        if (
-            event.action == Action.RIGHT_CLICK_BLOCK
-            && event.hand == EquipmentSlot.HAND
-            && event.useInteractedBlock() != Event.Result.DENY
-            && pylonBlock is PylonGuiBlock
-        ) {
-            event.setUseInteractedBlock(Event.Result.DENY)
-            event.setUseItemInHand(Event.Result.DENY)
-            val window = Window.single()
-                .setGui(pylonBlock.gui)
-                .setTitle(AdventureComponentWrapper(pylonBlock.name))
-                .setViewer(event.player)
-                .build()
-            window.open()
         }
     }
